remove get-apps graphql query
