+++ packages/backend/src/graphql/schema.graphql
@@ -1,9 +1,4 @@
 type Query {
-  getApps(
-    name: String
-    onlyWithTriggers: Boolean
-    onlyWithActions: Boolean
-  ): [App]
   getApp(key: String!, flowId: String): App
   getConnectedApps: [App]
   testConnection(
@@ -259,9 +254,9 @@ type App {
   docUrl: String
   authDocUrl: String
   primaryColor: String
-  auth: JSONObject
-  triggers: JSONObject
-  actions: JSONObject
+  auth: AppAuth
+  triggers: [Trigger]
+  actions: [Action]
   connections: [Connection]
   description: String
   isNewApp: Boolean
