+++ packages/frontend/src/graphql/queries/get-app.ts
@@ -9,7 +9,53 @@ export const GET_APP = gql`
       docUrl
       authDocUrl
       primaryColor
-      auth
+      auth {
+        connectionType
+        connectionRegistrationType
+        fields {
+          key
+          label
+          type
+          required
+          readOnly
+          value
+          description
+          docUrl
+          allowArbitrary
+          clickToCopy
+          autoComplete
+          options {
+            label
+            value
+          }
+        }
+        authenticationSteps {
+          type
+          name
+          arguments {
+            name
+            value
+            type
+            properties {
+              name
+              value
+            }
+          }
+        }
+        reconnectionSteps {
+          type
+          name
+          arguments {
+            name
+            value
+            type
+            properties {
+              name
+              value
+            }
+          }
+        }
+      }
     }
   }
 `
