+++ packages/frontend/src/graphql/queries/get-connected-apps.ts
@@ -10,7 +10,9 @@ export const GET_CONNECTED_APPS = gql`
       primaryColor
       connectionCount
       flowCount
-      auth
+      auth {
+        connectionType
+      }
     }
   }
 `
